+++ Classes/Domain/Model/JobPosition.php
@@ -101,6 +101,13 @@ class JobPosition extends AbstractEntity
     protected $requiredCommitments = '';
 
     /**
+     * benefits
+     *
+     * @var string
+     */
+    protected $benefits = '';
+
+    /**
      * workHours
      *
      * @var int
@@ -141,6 +148,13 @@ class JobPosition extends AbstractEntity
      * @var bool
      */
     protected $homeofficePossible = false;
+
+    /**
+     * directApply
+     *
+     * @var bool
+     */
+    protected $directApply = false;
 
     /**
      * currency
@@ -496,6 +510,26 @@ class JobPosition extends AbstractEntity
     }
 
     /**
+     * Returns the benefits
+     *
+     * @return string $benefits
+     */
+    public function getBenefits()
+    {
+        return $this->benefits;
+    }
+
+    /**
+     * Sets the benefits
+     *
+     * @param string $benefits
+     */
+    public function setBenefits(string $benefits)
+    {
+        $this->benefits = $benefits;
+    }
+
+    /**
      * Returns the workHours
      *
      * @return int $workHours
@@ -653,6 +687,36 @@ class JobPosition extends AbstractEntity
     public function setHomeofficePossible(bool $homeofficePossible)
     {
         $this->homeofficePossible = $homeofficePossible;
+    }
+
+    /**
+     * Returns the directApply
+     *
+     * @return bool $directApply
+     */
+    public function getDirectApply()
+    {
+        return $this->directApply;
+    }
+
+    /**
+     * Returns the boolean state of directApply
+     *
+     * @return bool
+     */
+    public function isDirectApply()
+    {
+        return $this->directApply;
+    }
+
+    /**
+     * Sets the directApply
+     *
+     * @param bool $directApply
+     */
+    public function setDirectApply(bool $directApply)
+    {
+        $this->directApply = $directApply;
     }
 
     /**