+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.6] - 2022-07-12
+
+### Added
+- Checkbox for `direct_apply` in jopPositions.
+- `benefits` in jopPositions.
+- Structured Data for `experienceRequirements`, `responsibilities`, `jobBenefits`, `jobLocation,Type` and `directApply`.
 ## [2.1.5] - 2022-06-30
 
 ### Fixed